Refactor processes controllers
