+++ app/controllers/gobierto_admin/gobierto_participation/processes/process_contribution_containers_controller.rb
@@ -42,7 +42,7 @@ module GobiertoAdmin
             redirect_to(
               edit_admin_participation_process_contribution_container_path(process_id: @process.id,
                                                                            id: @contribution_container_form.contribution_container.id),
-              notice: t(".success_html", link: gobierto_participation_process_process_contribution_container_url(@contribution_container_form.contribution_container.slug,
+              notice: t(".success_html", link: gobierto_participation_process_contribution_container_url(@contribution_container_form.contribution_container.slug,
                                                                                                                  process_id: @process.slug,
                                                                                                                  host: current_site.domain))
             )
@@ -64,7 +64,7 @@ module GobiertoAdmin
             redirect_to(
               edit_admin_participation_process_contribution_container_path(id: @contribution_container_form.contribution_container.id,
                                                                            process_id: @process.id),
-              notice: t(".success_html", link: gobierto_participation_process_process_contribution_container_url(@contribution_container_form.contribution_container.slug,
+              notice: t(".success_html", link: gobierto_participation_process_contribution_container_url(@contribution_container_form.contribution_container.slug,
                                                                                                                  process_id: @process.slug,
                                                                                                                  host: current_site.domain))
             )
