--- app/controllers/gobierto_participation/process_information_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module GobiertoParticipation
-  class ProcessInformationController < GobiertoParticipation::Processes::BaseController
-    def show
-      @information_text = current_process.information_text
-    end
-  end
-end
