+++ app/controllers/gobierto_participation/processes/information_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module GobiertoParticipation
+  module Processes
+    class InformationController < BaseController
+      def show
+        @information_text = current_process.information_text
+      end
+    end
+  end
+end
