+++ app/models/gobierto_cms/page.rb
@@ -53,18 +53,19 @@ module GobiertoCms
       collection.item_type.split('::').last.downcase
     end
 
+    # TODO: split methods to fetch news or pages
     def self.pages_in_collections(site)
-      ids = GobiertoCommon::CollectionItem.includes(:collection).where("collections.item_type = ? AND collection_items.item_type = ? ", %W(GobiertoCms::News GobiertoCms::Page), self.name).pluck(:item_id)
+      ids = GobiertoCommon::CollectionItem.where(item_type: %W(GobiertoCms::News GobiertoCms::Page)).pluck(:item_id)
       where(id: ids, site: site)
     end
 
     def self.pages_in_collections_and_container_type(site, container_type)
-      ids = GobiertoCommon::CollectionItem.includes(:collection).where("collections.item_type = ? AND collection_items.item_type = ? AND collection_items.container_type = ?", %W(GobiertoCms::News GobiertoCms::Page), self.name, container_type).pluck(:item_id)
+      ids = GobiertoCommon::CollectionItem.where(item_type: %W(GobiertoCms::News GobiertoCms::Page), container_type: container_type).pluck(:item_id)
       where(id: ids, site: site)
     end
 
     def self.pages_in_collections_and_container(site, container)
-      ids = GobiertoCommon::CollectionItem.includes(:collection).where("collections.item_type = ? AND collection_items.item_type = ? AND collection_items.container_type = ? AND collection_items.container_id = ?", %W(GobiertoCms::News GobiertoCms::Page), self.name, container.class, container.id).pluck(:item_id)
+      ids = GobiertoCommon::CollectionItem.where(item_type: %W(GobiertoCms::News GobiertoCms::Page), container_type: container_type, container_id: container.id).pluck(:item_id)
       where(id: ids, site: site)
     end
