+++ app/views/gobierto_participation/flags/create.js.erb
@@ -1,3 +1,3 @@
 <% if @flag %>
-  $("#flag-link-<%= @flag.flaggable.id %>").html("<%= j render(partial: 'gobierto_participation/process_contributions/flag', locals: { item: @flag.flaggable }) %>");
+  $("#flag-link-<%= @flag.flaggable.id %>").html("<%= j render(partial: 'gobierto_participation/processes/contributions/flag', locals: { item: @flag.flaggable }) %>");
 <% end %>
