+++ app/views/gobierto_participation/flags/destroy.js.erb
@@ -1,1 +1,1 @@
-$("#flag-link-<%= @flag.flaggable.id %>").html("<%= j render(partial: 'gobierto_participation/process_contributions/flag', locals: { item: @flag.flaggable }) %>");
+$("#flag-link-<%= @flag.flaggable.id %>").html("<%= j render(partial: 'gobierto_participation/processes/contributions/flag', locals: { item: @flag.flaggable }) %>");
