--- app/views/gobierto_participation/process_contributions/_comment_votes.html.erb
@@ -1,33 +1,0 @@
-<% if user_signed_in? %>
-  <div id="votes-links-<%= comment.id %>">
-    <div class="action">
-      <%= link_to gobierto_participation_process_process_contribution_container_process_contribution_vote_path(process_contribution_id: comment.commentable.slug,
-                                                                                                               process_contribution_container_id: comment.commentable.contribution_container.slug,
-                                                                                                               process_id: comment.commentable.contribution_container.process.slug,
-                                                                                                               vote_weight: 1,
-                                                                                                               votable_id: comment.id,
-                                                                                                               votable_type: comment.class),
-                  class: "action #{ 'active' if (comment.voted_with_weight_by_user?(current_user, 1))}",
-                  method: (comment.voted_with_weight_by_user?(current_user, 1) ? :delete : :post),
-                  remote: true do %>
-      <i class="fa fa-thumbs-up"></i>
-      <% end %>
-      <%= comment.positive_votes %>
-    </div>
-
-    <div class="action">
-      <%= link_to gobierto_participation_process_process_contribution_container_process_contribution_vote_path(process_contribution_id: comment.commentable.slug,
-                                                                                                               process_contribution_container_id: comment.commentable.contribution_container.slug,
-                                                                                                               process_id: comment.commentable.slug,
-                                                                                                               vote_weight: -1,
-                                                                                                               votable_id: comment.id,
-                                                                                                               votable_type: comment.class),
-                  class: "action #{ 'active' if (comment.voted_with_weight_by_user?(current_user, -1))}",
-                  method: (comment.voted_with_weight_by_user?(current_user, -1) ? :delete : :post),
-                  remote: true do %>
-      <i class="fa fa-thumbs-down"></i>
-      <% end %>
-      <%= comment.negative_votes %>
-    </div>
-  </div>
-<% end %>
