+++ app/views/gobierto_participation/processes/contributions/show.js.erb
