+++ app/views/gobierto_participation/processes/information/show.html.erb
@@ -24,7 +24,7 @@
   <div class="pure-g">
     <div class="pure-u-1 pure-u-lg-18-24">
       <article class="news_article">
-        <%= raw @information_text %>
+        <%== @information_text %>
       </article>
     </div>
