+++ app/views/gobierto_participation/processes/show.html.erb
@@ -61,7 +61,7 @@
         <h3 class="with_light_separator"><%= t('.about') %></h3>
         <p><%= current_process.body %></p>
 
-        <%= link_to t('.more_information'), gobierto_participation_process_process_information_path(process_id: @process.slug), class: "button outline rounded" %>
+        <%= link_to t('.more_information'), gobierto_participation_process_information_path(process_id: @process.slug), class: "button outline rounded" %>
       </div>
     </div>
