+++ app/views/gobierto_participation/votes/create.js.erb
@@ -1,5 +1,5 @@
 <% if @vote %>
   <% if @vote.votable_type == "GobiertoParticipation::Comment" %>
-    $("#votes-links-<%= @vote.votable.id %>").html("<%= j render(partial: 'gobierto_participation/process_contributions/comment_votes', locals: { comment: @vote.votable }) %>");
+    $("#votes-links-<%= @vote.votable.id %>").html("<%= j render(partial: 'gobierto_participation/processes/contributions/comment_votes', locals: { comment: @vote.votable }) %>");
   <% end %>
 <% end %>
