+++ app/views/gobierto_participation/votes/destroy.js.erb
@@ -1,1 +1,1 @@
-$("#votes-links-<%= @vote.votable.id %>").html("<%= j render(partial: 'gobierto_participation/process_contributions/comment_votes', locals: { comment: @vote.votable }) %>");
+$("#votes-links-<%= @vote.votable.id %>").html("<%= j render(partial: 'gobierto_participation/processes/contributions/comment_votes', locals: { comment: @vote.votable }) %>");
