+++ config/i18n-tasks.yml
@@ -43,25 +43,25 @@ ignore_unused:
   - 'helpers.page_entries_info.*'
   - 'gobierto_indicators.cards.*'
   - 'user.subscriptions.subscribable_button.*'
-  - 'gobierto_participation.process_contribution_containers.show.login_signup_idea'
-  - 'gobierto_participation.process_contribution_containers.show.login_signup_proposal'
-  - 'gobierto_participation.process_contribution_containers.show.login_signup_question'
-  - 'gobierto_participation.process_contribution_containers.show.put_idea'
-  - 'gobierto_participation.process_contribution_containers.show.put_proposal'
-  - 'gobierto_participation.process_contribution_containers.show.put_question'
-  - 'gobierto_participation.process_contribution_containers.show.show_ideas'
-  - 'gobierto_participation.process_contribution_containers.show.show_proposals'
-  - 'gobierto_participation.process_contribution_containers.show.show_questions'
-  - 'gobierto_participation.process_contributions.contribution.values_this_idea'
-  - 'gobierto_participation.process_contributions.contribution.values_this_proposal'
-  - 'gobierto_participation.process_contributions.contribution.values_this_question'
-  - 'gobierto_participation.process_contribution_containers.show.put_idea'
-  - 'gobierto_participation.process_contribution_containers.show.put_proposal'
-  - 'gobierto_participation.process_contribution_containers.show.put_question'
-  - 'gobierto_participation.process_contribution_containers.show.idea.*'
-  - 'gobierto_participation.process_contribution_containers.show.proposal.*'
-  - 'gobierto_participation.process_contribution_containers.show.question.*'
-  - 'gobierto_participation.process_contributions.form.*'
+  - 'gobierto_participation.processes.contribution_containers.show.login_signup_idea'
+  - 'gobierto_participation.processes.contribution_containers.show.login_signup_proposal'
+  - 'gobierto_participation.processes.contribution_containers.show.login_signup_question'
+  - 'gobierto_participation.processes.contribution_containers.show.put_idea'
+  - 'gobierto_participation.processes.contribution_containers.show.put_proposal'
+  - 'gobierto_participation.processes.contribution_containers.show.put_question'
+  - 'gobierto_participation.processes.contribution_containers.show.show_ideas'
+  - 'gobierto_participation.processes.contribution_containers.show.show_proposals'
+  - 'gobierto_participation.processes.contribution_containers.show.show_questions'
+  - 'gobierto_participation.processes.contributions.contribution.values_this_idea'
+  - 'gobierto_participation.processes.contributions.contribution.values_this_proposal'
+  - 'gobierto_participation.processes.contributions.contribution.values_this_question'
+  - 'gobierto_participation.processes.contribution_containers.show.put_idea'
+  - 'gobierto_participation.processes.contribution_containers.show.put_proposal'
+  - 'gobierto_participation.processes.contribution_containers.show.put_question'
+  - 'gobierto_participation.processes.contribution_containers.show.idea.*'
+  - 'gobierto_participation.processes.contribution_containers.show.proposal.*'
+  - 'gobierto_participation.processes.contribution_containers.show.question.*'
+  - 'gobierto_participation.processes.contributions.form.*'
 
 translation:
   api_key: ''
