+++ test/integration/gobierto_participation/processes/process_contribution_create_test.rb
@@ -17,14 +17,14 @@ module GobiertoParticipation
     end
 
     def container_path
-      @container_path ||= gobierto_participation_process_process_contribution_container_path(
+      @container_path ||= gobierto_participation_process_contribution_container_path(
         process_id: process.slug,
         id: contribution_container.slug
       )
     end
 
     def containers_path
-      @containers_path ||= gobierto_participation_process_process_contribution_containers_path(
+      @containers_path ||= gobierto_participation_process_contribution_containers_path(
         process_id: process.slug
       )
     end
